Refactor tests to use common stubs class.
